Update setup-director for newer stemcells
Previously, stemcells such as the 3421 series would run into the
following error during run-bats:
`Authentication failed with gateway 35.193.104.140 and user vcap`
One example which failed is
`network configuration resolving DNS entries resolves instance
names from deployed VM`
This was caused by the switch to not including the vcap user in
the `bosh_sshers` group.

We use the `user_add` job included in the `os-conf` release to
include the vcap user inside the `bosh_sshers` group. This also
explicitly sets the vcap user's authorized public key to be
compatible with the provided private key.
https://github.com/cloudfoundry/os-conf-release

Once part of this upgrade is switching to bosh-cli V2.
This moved us from the old style of `${}` templating to taking
advantage of the new cli's variable interpolation, `(())`.
Notably, we generate explicit certificates as the V2 cli
performs strict validation.
The aim of moving to the new cli is to eliminate the potential
for breakage that may occur when using the older cli.
https://github.com/cloudfoundry/bosh-cli/